pyproject.toml
[tool.poetry]
name = "chacha-chatbot"
version = "0.2.0"
description = ""
authors = ["Young-Ho Kim", "Woosuk Seo"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "app" },
    { include = "backend" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
openai = "^1.13.3"
nanoid = "^2.0.0"
python-dotenv = "^1.0.0"
pendulum = "^3.0.0"
jsonlines = "^3.1.0"
chatlib = {git = "https://github.com/yghokim/chatlib.git", branch="main"}
gunicorn = "^22.0.0"
google-generativeai = "^0.3.1"
transformers = ">=4.55.0,<5.0.0"
streamlit = "^1.39.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
